[NuGet] Fix exception when creating VB.NET NUnit project.
Partially fixed bug 22908 - Getting exception while creating
VBNet->NUnit Library Project
https://bugzilla.xamarin.com/show_bug.cgi?id=22908

The underlying problem is that the VB.NET project template does not
create a project since it errors when trying to set the project's
target framework to .NET 4.5 since that does not seem to be supported.

The fix applied here is just to make the NuGet addin handle a failed
project creation. Previously it assumed that there would be a project
created when the template defined NuGet packages for the project.
Now if there are no projects created an exception will not be thrown.

Another fix applied was to not try to install an empty list of package
actions which was causing the message "Packages successfully added" to
the status bar even though no NuGet packages were installed.