Hack to build against Mvc3 on tarball builds and Mvc4 on git builds.
If you're building from git, you get Mvc4 from NuGet, courtesy of an
unversioned assembly reference pointing to the NuGet path as a hint.
However, if this NuGet file does not exist, Xbuild resolves the Mvc
reference to Mvc1, even if 2 or 3 are available on the system (1-3 are
distributed with Mono). Here, we add a conditional reference to Mvc3
which is true on tarball builds - the second reference still resolves
to Mvc1, but is not used. On git builds, only the second unversioned
reference is used, which resolves to Mvc4 because the HintPath points
to that version, provided by NuGet.