Allow subclasses to modify loaded/saved text
Hypothetically, a subclass of SourceEditorView might want to store metadata
in a hidden header area of the file. Overriding ProcessLoadText gives the
subclass an opportunity to extract said metadata, and ensure that the text
loaded into the Document excludes it.

Similarly, ProcessSaveText would allow this subclass to inject updated
metadata into the file text just before the write occurs.