[NuGet] Handle recent packages that are removed from local file system.
Fixed bug 23177 - Unable to add NuGet package again after removing it.
https://bugzilla.xamarin.com/show_bug.cgi?id=23177

Adding a NuGet package to the project could fail with an error message
"Could not find a part of the path" if a package that exists in the
project is installed again. Installing the package again if it was
already installed in the project causes the local package to be added
to the most recently used package list. Then if the package was
uninstalled the recently used package list was still pointing to the
package on the local machine but it no longer exists. Now the recently
used package list only returns local packages that still exist on the
file system.