[Ide] Fix formatting document removing all XML text.
Fixed bug 20467 - Format document on duplicated attributes removes all text
https://bugzilla.xamarin.com/show_bug.cgi?id=20467

If the XML in the text editor was invalid when the Edit - Format -
Format Document menu was selected the text would be removed from the
text editor.

Now if the XML is invalid the text is not removed. An error is logged
but nothing else happens.

Whilst this could have been fixed in the XmlFormatter with it
returning the original XML unmodified doing so would have meant
that the undo history would change even though the file had not
been modified. The file would also have been shown as edited.