[Ide] Show project templates in New Project Dialog.
Added a new extension path /MonoDevelop/Ide/ProjectTemplatingProviders
which allows the loading of project templates to be extended.

All project templates are now shown in the New Project Dialog. Since
the current template categories do not match the new categories for
the dialog all the project templates are put in the default category
Other/Miscellaneous/General.