[Ide] Show valid directory names in new project preview.
Directory names for the solution and project have any path or volume
separators removed that may have been typed into the New Project
dialog for the project name or solution name. The preview of the
folder structure now shows the actual directory names that would be
created for the new project.