[Ide] Update language selector behaviour in New Project Dialog.
If a project template has no language (e.g. Blank Solution template)
then no language selector is displayed when the template is selected.

If a project template has one language then a language selector is
displayed without a drop down arrow when the template is selected.
Also the language text is centered in the language selector.

If a project template has two or more languages then a language
selector is displayed with a drop down arrow.

If the currently selected language does not have any corresponding
languages in a template when it is selected then the first language
will be displayed in the language selector.