[Ide] Support grouping project templates.
If a project template has a _GroupId element as part of its
TemplateConfiguration then this will be used by the New Project Dialog
to group the project templates together.

    <_GroupId>md-project-console</_GroupId>

In the New Project dialog all project templates with the same group id
are shown as one row in the dialog if they exist in the same category.
To select a different project template the language selector drop down
should be used. On selecting a different template the right hand side
of the dialog, showing the detailed template description, will be
updated with the information for that template.