[Ide] Always show language for template in New Project Dialog.
Previously the language selector would only be displayed if a project
template row was selected in the template list and the row was active.
Tabbing around the dialog would then hide the language of the project
template.

Now the language selector is shown at all times when a project
template is selected even if the row does not have the focus.