[Ide] Implement browse button in New Project dialog.
The Browse button can now be used to select a folder for the new
project.

The New Dialog is not using the FolderEntry widget and uses its own
text box and button since the layout for these is driven by a table
and they are both in different columns. The FolderEntry widget has its
own HBox containing a text box and a button.