[Ide] Prevent selection of top level categories in New Project dialog.
The top level categories on the left hand side of the New Project
dialog can no longer be selected:

   Cross-Platform
   iOS
   Android
   OS X
   Other

Previously these could be selected and but no templates were
displayed. Other IDEs, such as XCode, allow you to select the top
level categories and will display all the templates in that category.
For now the top level categories cannot be selected.