[NuGet] Handle failure to read NuGet.Config
Fixed bug 23436 - Solution pad causes errors and fails to show
https://bugzilla.xamarin.com/show_bug.cgi?id=23436

If the ~/.config/NuGet directory cannot be created by NuGet then an
unhandled UnauthorizedAccessException is thrown. This prevents the
solution pad from working. Now if there is any error creating this
directory or trying to load the NuGet.Config file then the error is
logged and a NuGet.NullSettings is returned. This allows the IDE
to continue  working. If the ~/.config/NuGet directory cannot be
created by NuGet then it will not be possible to use NuGet in the
IDE but it will not prevent the solution pad from being used.