[Ide] Hide solution templates when adding to an existing solution.
The original New Project dialog would not show templates that would
only create a solution but not a project when adding a new project to
an existing solution. This means the Workspace template and the
Blank Solution templates are not shown when adding a new project to
an existing solution.