[Ide] Update solution name with project name in New Project dialog.
When creating a new solution the New Project dialog will now use the
project name as the solution name by default. When the project name is
typed in the solution name text box will be filled in with the project
name. If the solution name is changed so it is different to the the
project name then the solution name text box will no longer be
updated as the project name is changed.