[Ide] Use correct project file extension in New Project dialog preview
The final configuration page in the New Project dialog now shows the
correct project file extension. The project file extension will be
taken from the project template if it exists. The project file
template now has an optional _FileExtension element which can be used

    <_FileExtension>.myproj</_FileExtension>

If there is no file extension defined for the template then the
project file extension will be determined based on the language name
defined in the project template. Currently only C#, VBNet and IL
languages are recognised. If the file extension cannot be determined
then .proj will be used by default.