[Ide] Improve workspace support in New Project dialog.
The workspace template is treated as a special case in the final
configuration page in the New Project dialog:

"Workspace Name" label used instead of "Solution Name".
Folder preview uses Workspace.mdw by default.
Banner text at the top of the dialog now refers to Workspace instead
of Project.

When a solution template is selected that has no projects the banner
text at the top of the dialog now refers to Solution instead of
Project.