[Ide] Avoid enabling a toolbar and a pathbar at the same time in a SdiWorkspaceWindow context.
The Android designer makes use of a main view (AndroidDesignerView) and a subview (LayoutSourceEditorView). AndroidDesignerView overrides the GetContent method to proxy some calls to that editor subview. This notably allow Document.InitializeExtensionChain to be called correctly and create/register the autocompletion extension specific to Android (AndroidXmlEditorExtension).

The main issue is that due to the way Document's GetContent works (also querying the extension points), our initial AndroidDesignerView.GetContent override means that, by a kind of misfortune, our Android XML extension is picked up as a IPathedDocument source by Document.cs:665 (code below):

    ((SdiWorkspaceWindow)window).AttachToPathedDocument (GetContent<MonoDevelop.Ide.Gui.Content.IPathedDocument> ());

This causes SdiWorkspaceWindow to automatically create a PathBar in its AttachToPathedDocument method. Problem is that AndroidDesignerView also manually create a Toolbar (via WorkbenchWindow.GetToolbar) and both bar widgets seem to have never been made to work together. The result is that when both are shown, they badly overlap (the Allocation of the Toolbar is set 22px down but the drawing code doesn't take it into account).

This patch mitigates the issue by tracking toolbar state before enabling a pathbar. It also tracks subview content changes to re-enable the pathbar when it makes sense by checking that the toolbar is disabled in that case.