[Ide] Change New Project dialog size and background colours.
Update the New Project dialogs height and width to match the UI
design. The height may need to increase to fit the categories
when padding is added to the top of the tree view.
Change the first and second column widths to be smaller.
Change the background colour of the first and second columns.
Change case of Cross-platform category title to match UI design.