[Ide] Custom draw each row in the project template list.
Custom draw the icon and template name text in the project template
list (second column) in the New Project dialog.
This allows more control of the layout of the icon and text so the
dialog can look like the UI design.
Increased the height of the template category row.

If the template name is too long for the column then an ellipsis is
drawn at the end of the text. If the language selector is displayed
then the text is no longer hidden behind the language selector and
an ellipsis is used at the end of the text if necessary.