[NuGet] Fix types from MonoGame NuGet package not being recognised.
After adding the MonoGame.Binaries NuGet package the type system does
not recognise any types that were imported by the custom MSBuild
target file. The types are shown in red text in the text editor. To
fix this the solution had to be closed and re-opened.

This is related to bug 23626 - Xamarin Studio doesn't recognize that
files exist after NuGet Update
https://bugzilla.xamarin.com/show_bug.cgi?id=23626

The MonoGame.Binaries NuGet package has a custom MSBuild .targets file
that adds extra references:

<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
  <ItemGroup>
    <Reference Include="MonoGame.Framework">
      <HintPath>$(MSBuildThisFileDirectory)\MonoGame.Framework.dll</HintPath>
    </Reference>
    <Reference Include="Tao.Sdl">
      <HintPath>$(MSBuildThisFileDirectory)\Tao.Sdl.dll</HintPath>
    </Reference>
    <Reference Include="OpenTK">
      <HintPath>$(MSBuildThisFileDirectory)\OpenTK.dll</HintPath>
    </Reference>
  </ItemGroup>
</Project>

Now after a NuGet package is installed and it contains MSBuild targets
file then the type system will be refreshed for that project. The
types will then be known by Xamarin Studio and no longer show as red
text in the text editor.