[Ide] Fix New Project dialog zero border width
Change the name of the dialog to wizard_dialog so the
Dialog.VBox.BorderWidth is set to zero for the content area.
Previously the only way to get this to work was to set the
BorderWidth after Dialog.Show () is called. Unfortunately this
has the side effect that the border is shown briefly and then
the content of the dialog expands to fill the gap. All of
which can be seen in the dialog as it opens.

Currently only works on the Mac.