[Ide] Increase width for template list in New Project Dialog.
The second column showing the list of templates now has increased
width to allow the template names to be displayed without being
truncated.

The third column with the template description now has a smaller image
and more room for the description. The template name is now wrapped.
The image is resized in code and needs replacing with one that is
of the correct size.

Adjusted the preview widget so its right hand column is the same width
as the template description column.