[Ide] Fix project template description label moving down dialog
When opening the New Project dialog for the first time the
long template description would move down to make room for
the template heading just above it. This movement was
visible on opening the dialog. To workaround this the
label is given a default bit of text that is marked up
so GTK can determine the size required for the label. This
prevents the long description label from being moved down
when the dialog is first opened. You can still see the
long description label being wrapped after the dialog is
opened for the first time. A possible workaround here is
to set the initial text to match the first template that
will be selected so the wrapping is completed before the
label is visible.