[Ide] Support conditionally creating projects in project template.
Project templates can now choose whether to create a project or not
based on parameters defined by the template wizard. The wizard can
set a property to be true or false and then the project template will
check the property is true against the condition defined in the
template. The project template now has an "if" attribute with the
property that will be checked.

   <Project
        name = "${ProjectName}"
        directory = "."
        if = "CreateProject">