[Ide] Do not add project that fails condition to solution.
When adding a project to an existing solution only the first project
in the project template is currently added to the solution. If the
project template has conditions then the wrong project can be added
to the solution. Now the condition is checked so the first project
that passes the condition is added to the solution.