[Ide] Project templates can now define supported wizard parameters.
In the project template the parameters that a template wizard supports
is defined in the _SupportParameters element.

  <_SupportedParameters>AppName,AppIdentifier</_SupportedParameters>

This information is supplied to the template wizard which can then
show or hide UI controls based on the supported parameters.

If the template has no supported parameters defined then the template
wizard will assume all parameters are supported.