[Ide] Fix Git option in New Project dialog.
Changing the "Use git for version control" check box was not disabling
the creation of a Git repository when creating a new solution.

If the "Use git for version control" check box is unchecked then the
"Create .gitignore file" check box is disabled.

Project structure preview now hides the .gitignore file if the "Use
git" check box is unchecked.