[Ide] Remember create project directory setting in New Project dialog.
The "Create project directory within the solution directory" check box
state is saved when a new solution or project is created. Two
properties are used depending on whether a new solution is being
created or a new project is created so the check box state is saved
separately depending on the context. The main reason for a different
saved setting when adding a new project to an existing solution is
that if it was shared and you uncheck the option when creating a new
solution then adding a new project to the solution would try to add
the project directly into the solution directory which generally is
not what you want.