[Refactoring] Enhance GotoDeclaration to support multiple open solutions.
If you have more than one open solution, using GotoDeclaration for a type in the other, non-active solution, results in that type being displayed in the Assembly Browser. This change changes that behaviour to search for the type in any of the other open solutions and open the source file for the type instead.