[Core] Fix project building issue
The project builder is not more aggressive when a project needs to be
unloaded. Instead of just unloading the project element, it now
unloads all projects. This is necessary because the project may reference
other target files that may have been changed, and which can't be
individually unloaded.

Fixes bug #23435 - XS Shared Projects do not work unless you readd reference