[Ide] Cache and dispose wizard controls.
Wizard controls are cached whilst moving back and forth between the
pages. If the first page of the dialog is shown with the templates
then the wizard controls are disposed and recreated. If the Create
button is clicked then the wizard controls that were displayed are
disposed at that point.

Added unit tests for the wizard provider.