[Ide] Allow template to add file without file extension
Fixed bug 24143 - Default file extension is forced for RawFile when
creating a template
https://bugzilla.xamarin.com/show_bug.cgi?id=24143

Using <RawFile name=Icon src=Icon /> and specifying a file
without a file extension would create a file with an extension
taken from the language of the template. So if creating a project
from a C# template the file would have a .cs file extension
added to it.

Now if an empty DefaultExtension attribute is specified then a
default file extension will not be added to the file created.

<RawFile name=Icon src=Icon DefaultExtension= />

A default file extension will only be added if the file has
no DefaultExtension defined and the file created has no file
extension.