[Ide] Fix incorrect startup project on creating new solution.
Creating a new solution was always using the first project in the
solution as the startup project instead of the one defined in the
project template. When the startup project name defined in the
template was using place holder tags (e.g. ${ProjectName}) then these
were not being expanded so the startup project would never be found.