[SourceEditor] Introduce a new callback when about to set caret information.
When caret is trying to be updated externally (like when opening a document to a specified line/column), the existing caret callbacks are only executed when the caret has actually visually moved.

This causes an issue for view content with multiple subviews because the actual caret position update processing is linked to the SizeAllocated event (in TextArea) meaning that if the editor is in a subview that is not the one currently shown it will never be ran.

For view content in that situation (Android designer is an example), we need that extra callback to make sure we first switch the current subview to be the editor so that the rest of the chain is correctly processed.