[Ide] Prevent use of extra invalid characters when creating a project
Fixed bug 19137 - Semicolon should be a restricted character in
project name
https://bugzilla.xamarin.com/show_bug.cgi?id=19137

Using certain characters such as semicolon, asterisk and ampersand
can cause build errors when they are part of the project or solution
name. So these three characters are now checked for on project
creation in the New Project dialog and will prevent the project from
being created.