[NuGet] Fix pre-release NuGet package being downgraded on update.
Fixed bug 24840 - On updating nuget package, Xamarin.Forms downgrades
from 1.3.0.6275-pre1 to 1.2.3.6257 and On building the solution
getting build error.
https://bugzilla.xamarin.com/show_bug.cgi?id=24840

When a pre-release NuGet package was installed that was newer than the
latest stable NuGet package then an Update would install the stable
version even though it was a lower version. Now an explicit check is
made to ensure that an older NuGet package is not being installed.