[Core] Prevent namespace ending with dot when creating new project.
Fixed bug 25108 - We should handle project name like "xxx.0"
https://bugzilla.xamarin.com/show_bug.cgi?id=25108

Creating a new project with a project name of "abc.1" will create a
new project with an invalid namespace of "abc.", since the digit is
removed, and the project will not compile.

Now when sanitizing the project's root namespace if it would end with
a '.' the namespace is corrected and made valid. So a project named
"abc.1" will be created with a namespace of "abc"