[Core] Add Unified Mac and iOS projects to known project types.
Fixed bug 24459 - Xamarin Studio or MonoDevelop throws exception
MonoDevelop.Projects.Formats.MSBuild.UnknownSolutionItemTypeException
on adding iOS Unified project to existing solution.
https://bugzilla.xamarin.com/show_bug.cgi?id=24459

Adding the project type guids for Unified Mac and iOS to the list of
known project types allows Unified Mac and iOS projects to be added
to an existing solution on Windows without an error being displayed.
Instead a warning icon will appear next to the project with a tooltip
indicating that the project is not supported in Xamarin Studio or
MonoDevelop on Windows.