[Ide] Change minimum size of the window.
1000 is just some arbitrary value, and it's too big to allow Windows
users to snap MonoDevelop windows left/right. So I'm choosing new
arbitrary sizes here. The new minimum width is 800, which means we'll
be able to snap left/right for resolutions as small as 1600px wide.
If we go much smaller than this, the toolbar is totally non-functional.
The new height is 540px, which is chosen because it's half of
1920x1080.

Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=22146