[NuGet] Fix build task not updated for Xamarin.Forms.
Fixed bug 23993 - Build task not updated after updating NuGet package
for Xamarin.Forms
https://bugzilla.xamarin.com/show_bug.cgi?id=23993

If a project that has a reference to Xamarin.Forms is built once and
then the NuGet package is updated the old MSBuild task dll is still
used when compiling after the update.

To reproduce this bug:

1) Create a Xamarin.Forms project and install Xamarin.Forms.1.0.6186
2) In the PCL project add a new content page so the Xaml Generator
is called on building the project.
3) Rebuild the PCL project. Compiling should fail with an error:
"XamlG Task failed unexpectedly"
4) Update the Xamarin.Forms NuGet package in the PCL project to the
latest version.
5) Rebuild the PCL project again. The "XamlG Task failed unexpectedly"
error is shown again. Also you can see in the Package Console that the
NuGet addin could not access the Xamarin.Forms.Build.Tasks.dll since
the Package Console reports the warning: "Access to the path
'Xamarin.Forms.Build.Tasks.dll' is denied."

The problem is that the MonoDevelop.Projects.Formats.MSBuild.exe that
compiles the project has locked the MSBuild task dll.

Now when an MSBuild import is removed on updating a NuGet package the
NuGet addin will dispose the current project builder for the project
which will kill the MonoDevelop.Projects.Formats.MSBuild.exe process.
This unlocks any custom MSBuild task dlls locked by the process and
allows the old NuGet package to be removed without any access denied
errors and also means when the project is recompiled again after the
update it will use the correct MSBuild task dll.