[Ide] Fix New button incorrectly disabled in New File Dialog.
Fixed bug 13645 - New button get disable for Empty Class when user
check/uncheck 'Add to Project' option in New File window.
https://bugzilla.xamarin.com/show_bug.cgi?id=13645

With a solution selected when the New File dialog is opened it has an
'Add to project' check box so you can add a file to a particular
project. With a file template selected toggling this 'Add to project'
check box would sometimes disable the New button when it should have
been enabled.

Now after the 'Add to project' check box is changed the New File
dialog will update the status of the New button so it is correctly
enabled or disabled based on the selected template and the values
entered in the New File dialog.