[NuGet] Handle exception when zero size .nupkg in packages directory.
Fixed bug 23759 - Package manager crashes when zero byte .nupkg is
in solution packages directory.
https://bugzilla.xamarin.com/show_bug.cgi?id=23759

Now if there is an exception when trying to get the solution packages
the error is logged and the solution packages are not displayed in the
Add Packages dialog. An error will be displayed in the Add Packages
dialog at the top of the dialog about being unable to read the zero
byte .nupkg file and any attempt to add a new NuGet package will fail
with the same error which will still be shown at the top of the
dialog.