[NuGet] Report correct error when checking for package updates.
If there is a zero byte .nupkg file in the solution's package
directory and check for updates is enabled then the error message
shown in the Package Console was:

"An exception was thrown while dispatching a method call in the UI
thread."

The DispatchService wraps the exception with its own so if a call is
dispatched to the UI thread the underlying error message is lost if
you use ex.Message and not ex.ToString (), which will show the full
call stack.

Now the NuGet addin will check if the exception thrown, whilst
checking for updates, has an InnerException and if the error message
is the one thrown by the DispatchService and will show the inner
exception's message instead in the Package Console.