[Ide] Fix warning creating solution with different name to project.
Fixed bug 23517 - Reference to unknown project when creating
Xamarin.Forms project with different solution name.
https://bugzilla.xamarin.com/show_bug.cgi?id=23517

When a project template has multiple projects one of which then has
a reference to another and the project name and solution name are
given different names in the New Project then a warning message
is shown saying 'Reference to unknown project'. The solution and
projects are created but the project will not have a reference to the
other project.

If a project template needs the original project name defined in the
New Project dialog's final configuration page there is a new template
property defined called UserDefinedProjectName which has this value.