[NuGet] Allow ASP.NET project templates to work offline.
With all the NuGet packages needed for the ASP.NET project templates
available in the local machine's cache it was not possible to create
one of these project templates without the NuGet packages failing to
install. This problem does not affect project templates, such as
Xamarin.Forms, which includes the NuGet packages alongside its project
templates.

Now the local machine's NuGet cache is used as the primary source of
packages for project templates so it is possible to create an
ASP.NET project template offline if the NuGet packages are in the
cache.