[NuGet] Fix updates available shown when packages have been updated.
Fixed bug 25505 - Updating all packages in a solution does not refresh
the update info in the Solution window.
https://bugzilla.xamarin.com/show_bug.cgi?id=25505

With check for updates enabled in Preferences, when an ASP.NET project
is created, updates are shown as available for several packages. When
the packages are updated for the solution or for the project the
Solution window still shows updates as being available even though
the packages have been updated.

The problem is that a change was made to the NuGet addin to update
a package dependencies at the same time as the package, also an extra
check is made to ensure the update is not trying to downgrade the
NuGet package. This meant that an update event was not fired for each
package being updated which results in some package updates being
shown as available in the Solution window even when they have been
updated.

Now the NuGet addin will check all package references when a package
is updated so if any package dependencies are updated the Solution
window will be updated to reflect the current status of the packages.