Added a FSharpFoldingParser implementation
The initial load of an F# document causes either an
IFoldingParser.Parse or a FSharpParser.Parse to be invoked.
FSharpParser.Parse uses the language service to make a blocking call to
retrieve the ParseAndCheckResults which
is usually no issue as this is on a background thread, however, during
load this is carried out on the UI thread.

Adding the default implementation of an IFoldingParser means that we
can return control to the UI quickly and still
have the async parse continue in the background which will update the
UI when completed.