Unload domain every time when number of refs get 0
If we use T4 Template which references assemblies generated by the project,
we need to load new version of assembly into templating domain after each 
compilation. To do it we must unload domain and create it again with new assemblies,
becase if domain holds an old version of the same assembly it does not allow to load
the assembly with the same name into it. 'unusedCount' field prevented to unload domain 
and T4 templates referenced the assemblies were got by the first compilation. 
This commit fixes this.