[NuGet] Fix second Package Console opened when creating new solution.
With check for package updates enabled in Preferences a second Package
Console was opened when a new solution was created from a project
template that added NuGet packages, such as the ASP.NET MVC project
template. This bug was introduced when the check for updates was
changed to run on its own separate thread. This separate thread would
create its own Package Console window to monitor progress before the
original progress monitor that was monitoring NuGet packages being
added to the project could be disposed.

Now the second Package Console window is not opened when the check
for updates runs after a new solution is created. One side effect is
that the Package Console is cleared when the check for updates
begins. However if there is an error when installing the NuGet
packages the check for updates will not be run so the error will
remain available in the Package Console.