Fixes for semantic highlighting
Fix for #if not been highlighted correctly
Fix for (*) not being escaped
Cleaned up code slightly, span parser seems overly complex.  It would
be preferable to use the base SpanParser (which uses the regex rules)
and just match up our language AST to the remaining plain text tokens,
leaving the preprocessor and comments formatted via the base
SpanParser.

The way it currently works it a mixture of custom span parsing and the
base span parser, this leads to discrepancies between the spans for
comments in the base class and the ones produced via the F# tokenizer.
Like seen in the numerous bugzilla issues raised against this component