[Core] Added ContextMenu abstractions that build the context menu UI at the time of the menu being shown.
Largely copied from Xwt.Menu but because Xwt decides on the backend at time of creation, and it defaults to Gtk, we don't get chance to use Mac native menus.
This will allow us to construct custom context menus (that are not based on command extensions) with arbitrary click handlers that will use the appropriate UI backend when shown.