[Ide] Fix null reference when adding multiple projects to solution.
Adding the projects from the ASP.NET MVC Project with Unit Tests
template to an existing solution would show an error: "Save failed.
Object reference not set to an instance of an object.". The new
projects however would still be added to the solution.

The problem was that the projects are saved before they are added to
a solution. This is a requirement of the version control system.
Without a parent solution when the unit tests project is saved a
null reference exception is thrown when it tries to resolve the
reference to the main MVC project being tested.

Note there are still problems with the namespace used in the
tests project created since it uses the solution name as the
namespace and not the project name.