Fixed unit tests
NUnit has support for tests that use async/await, however it doesn't work
very well with "async void" methods. NUnit needs to know that a method
is async, so when it is, it will wait for the spawned tasks to finish. When
the method is "async void", there is no task to wait on, so it does a trick:
it registers a new synchronization context that will record all async calls.
When the test ends, it will wait until all async calls recorded by the context
have finished. However, this doesn't work for MD's unit tests since MD
sets its own synchronization context. The solution is to avoid async void
test methods and use async test methods that return a task.