[Ide] Pass project string tag model to file templates
Previously we were priming SingleFileDescriptionTemplates's
custom tags with the ProjectCreationParameters.

Generalize this to passing file templates a project tag model.
For the case of new projects, this is the ProjectCreationParameters.
However, when adding files to an existing project, it's the
project's tag model.

Also fix project tags in file template references.