[Core] Refactored project creation
Changed the way creation data is provided to the constructor of
Project. Instead of using a global object, which was causing deadlocks
when used from different threads, we now use a SetCreationContext
method, which is called after creating the project. Since this data
needs to be assigned before the object is initialized, before creating
the project we set a logical context property which delays the
initialization. The object is later initialized using EnsureInitialized.