[Core] Fix several item metadata issues
The BuildItem class doesn't have a method to enumerate the metadata of the
item. Since we need that list in order to transfer the metadata from the
BuildItem to the MSBuildItem, we use a trick: we find the source of the
evaluated BuildItem and get the list from the xml.