[Ide] Project templates can now be hidden in the New Project dialog.
A project template can now indicate when it should be displayed in the
New Project dialog. The project template can indicate that it should
only be shown when creating a new solution, or when adding a new
project to an existing solution, or, by default, always be shown. This
is done by adding a new Visibility element to the template's header:

<TemplateConfiguration>
    <Visibility>NewProject</Visibility>

If this element is missing then the project template is always shown.
Supported Visibility values are: NewSolution and NewProject.